requirements.txt
# Flask Core
Flask==3.0.0
Werkzeug==3.0.1

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23

# Authentication & Security
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0

# Password hashing
bcrypt==4.1.2

# AI Integration
openai>=1.12.0
anthropic>=0.18.0
httpx>=0.24.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Development
Flask-Migrate==4.0.5
